Move pylama to tox
Tox is the go-to tool for our testing process, including linting. By moving
pylama to it, the only development dependency needed is `tox`, it'll take care
of creating the separate environments. Considering that the user already have
`tox` installed, the only thing she needs to do to run our entire test
suite (which includes linting) is:

```
tox
```